Fix update script.

+++ ha-loop.sh
@@ -10,21 +10,22 @@ if [[ -z "${POD_IP-}" ]]; then
   exit
 fi
 
-update_remote() {
-  local remote_ip="${1}"
+mapfile -t remote_ips < <(ha_remote_addresses)
 
-  local current_remote_ip
+if [[ ${#remote_ips[@]} -ge 1 ]]; then
+  remote_ip="${remote_ips[0]}"
   current_remote_ip="$(ha_conf_get remote)"
 
   if [[ "${remote_ip}" != "${current_remote_ip}" ]]; then
+    if [[ ${#remote_ips[@]} -ge 2 ]]; then
+      echo "Warning: Multiple remotes found (${remote_ips}), choosing first."
+    fi
+
     echo "Updating high availability remote to '${remote_ip}'."
     ha_conf_set remote "${remote_ip}"
-
-    sed -i -E "s/#\s+(remote\s+=).*/\1 ${1}/" /etc/strongswan.d/charon/ha.conf
     kill -HUP "${server_pid}"
   fi
-}
-
-if remote_ip="$(ha_remote_addresses | head -n 1)"; then
-  update_remote "${remote_ip}"
 fi
+
+heartbeat_delay="$(ha_conf_get heartbeat_delay)"
+sleep "${heartbeat_delay}e-3"
